Remove Beaker checkServer check for /permabit/release.
/permabit/release was removed from Beaker mounts as an always present
mount due to the issue that use of binary artifacts would be attempted
and this conflicted with architectures other than x86_64.  The check
for /permabit/release in checkServer should have been removed when
this change was made.

The use of /permabit/release is specific to upgrade/cross-platform
testing and these tests specify the mount of /permabit/release.

Pair: slegendr

.../main/src/tools/bin/checkServer.pl
Removed check for /permabit/release in Beaker.

.../common/perl/bin/checkServer.pl
Removed check for /permabit/release in Beaker.

.../perl-Permabit-checkServer.spec
Updated release and changelog.

[git-p4: depot-paths = "//eng/common/": change = 157305]